pyproject.toml
[project]
name = "ia-tarea3"
version = "0.1.0"
description = ""
authors = [
    {name = "bitfalt"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain-openai (>=0.3.22,<0.4.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
